Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
aiosmtplib
email-validator
slowapi
python-dotenv